chore: split analysis modules
