+++ src/gem/analysis/__init__.py
@@ -1,0 +1,44 @@
+"""Post-parse analysis helpers for gem replay data."""
+
+from gem.analysis.abilities import ability_level_at_tick
+from gem.analysis.combat import (
+    AbilityCast,
+    group_ability_hits,
+    is_active_teamfight_participant,
+    teamfight_at_tick,
+)
+from gem.analysis.formatting import format_npc_name
+from gem.analysis.map_context import (
+    CampVisitContext,
+    MapContextBucket,
+    build_map_context_timeline,
+    score_camp_visit_context,
+    world_in_bounds,
+)
+from gem.analysis.roshan import RoshConversion, RoshTimelineEvent, build_rosh_conversions
+from gem.analysis.spatial import heroes_near, net_worth_at, position_at_tick
+from gem.analysis.vision import VisionSource, _is_daytime, estimate_vision, ward_vision_impact
+
+__all__ = [
+    "AbilityCast",
+    "CampVisitContext",
+    "MapContextBucket",
+    "RoshConversion",
+    "RoshTimelineEvent",
+    "VisionSource",
+    "_is_daytime",
+    "ability_level_at_tick",
+    "build_map_context_timeline",
+    "build_rosh_conversions",
+    "estimate_vision",
+    "format_npc_name",
+    "group_ability_hits",
+    "heroes_near",
+    "is_active_teamfight_participant",
+    "net_worth_at",
+    "position_at_tick",
+    "score_camp_visit_context",
+    "teamfight_at_tick",
+    "ward_vision_impact",
+    "world_in_bounds",
+]
